requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==4.2.19
djangorestframework==3.15.2
flake8==7.1.2
gunicorn==23.0.0
idna==3.10
mccabe==0.7.0
packaging==24.2
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
stripe==11.5.0
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
